backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
acrcloud
ffmpeg-python
pydub
google-generativeai
requests
python-dotenv
